requirements.txt
python-dotenv==1.0.1
confluent-kafka==2.5.3